refactor: finish DocsTab and ws-handler responsibility split
Extract remaining cohesive units guided by moniker callable clusters:
DocsTab shell vs toolbar/category/selection/bulk/item modules, and
ws-handler domain command packs (document, charte, asset, text).
